WorkspaceMenuComponentを追加 #42
